+++ task/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.1
 
 - Fix issue where connection used same PEM files after removing them
+- Improve feedback for incorrect taskd.credentials
 - Display start field
 
 # 0.2.0